reorganize crates in preparation for split
