+++ filecoin-proofs/src/api/constants.rs
@@ -1,0 +1,2 @@
+pub const POST_SECTORS_COUNT: usize = 2;
+pub const POREP_MINIMUM_CHALLENGES: usize = 12; // FIXME: 8,000
