+++ filecoin-proofs/src/api/internal.rs
@@ -1,20 +1,12 @@
-use std::collections::HashMap;
-use std::fs::{copy, remove_file, File, OpenOptions};
+use std::fs::{copy, File, OpenOptions};
 use std::io::{BufWriter, Read};
 use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
 
-use bellperson::groth16;
 use ff::PrimeField;
-use fil_sapling_crypto::jubjub::JubjubBls12;
 use memmap::MmapOptions;
-use paired::bls12_381::{Bls12, Fr};
+use paired::bls12_381::Bls12;
 use paired::Engine;
 
-use crate::api::post_adapter::*;
-use crate::error;
-use crate::error::ExpectWithBacktrace;
-use crate::FCP_LOG;
 use sector_base::api::bytes_amount::{PaddedBytesAmount, UnpaddedByteIndex, UnpaddedBytesAmount};
 use sector_base::api::porep_config::PoRepConfig;
 use sector_base::api::porep_proof_partitions::PoRepProofPartitions;
@@ -23,301 +15,291 @@ use sector_base::api::post_proof_partitions::PoStProofPartitions;
 use sector_base::api::SINGLE_PARTITION_PROOF_LEN;
 use sector_base::io::fr32::write_unpadded;
 use storage_proofs::circuit::multi_proof::MultiProof;
-use storage_proofs::circuit::vdf_post::{VDFPoStCircuit, VDFPostCompound};
+use storage_proofs::circuit::vdf_post::VDFPostCompound;
 use storage_proofs::circuit::zigzag::ZigZagCompound;
 use storage_proofs::compound_proof::{self, CompoundProof};
-use storage_proofs::drgporep::DrgParams;
 use storage_proofs::drgraph::{DefaultTreeHasher, Graph};
 use storage_proofs::fr32::{bytes_into_fr, fr_into_bytes, Fr32Ary};
 use storage_proofs::hasher::pedersen::{PedersenDomain, PedersenHasher};
 use storage_proofs::hasher::{Domain, Hasher};
-use storage_proofs::layered_drgporep::{self, ChallengeRequirements, LayerChallenges};
+use storage_proofs::layered_drgporep::{self, ChallengeRequirements};
 use storage_proofs::merkle::MerkleTree;
 use storage_proofs::porep::{replica_id, PoRep, Tau};
-use storage_proofs::proof::{NoRequirements, ProofScheme};
-use storage_proofs::vdf_post::{self, VDFPoSt};
-use storage_proofs::vdf_sloth::{self, Sloth};
+use storage_proofs::proof::NoRequirements;
+use storage_proofs::vdf_post;
+use storage_proofs::vdf_sloth;
 use storage_proofs::zigzag_drgporep::ZigZagDrgPoRep;
-use storage_proofs::zigzag_graph::ZigZagBucketGraph;
 
-pub type Commitment = Fr32Ary;
-pub type ChallengeSeed = Fr32Ary;
+use crate::api::caches::{
+    get_post_params, get_post_verifying_key, get_zigzag_params, get_zigzag_verifying_key,
+};
+use crate::api::constants::POREP_MINIMUM_CHALLENGES;
+use crate::api::file_cleanup::FileCleanup;
+use crate::api::parameters::{post_setup_params, public_params, setup_params};
+use crate::api::post_adapter::*;
+use crate::api::singletons::ENGINE_PARAMS;
+use crate::error;
+use crate::error::ExpectWithBacktrace;
+use crate::FCP_LOG;
 
 /// FrSafe is an array of the largest whole number of bytes guaranteed not to overflow the field.
 type FrSafe = [u8; 31];
 
-lazy_static! {
-    pub static ref ENGINE_PARAMS: JubjubBls12 = JubjubBls12::new();
-}
+pub type Commitment = Fr32Ary;
+pub type ChallengeSeed = Fr32Ary;
+type Tree = MerkleTree<PedersenDomain, <PedersenHasher as Hasher>::Function>;
 
-////////////////////////////////////////////////////////////////////////////////
-/// Groth Params/Verifying-keys Memory Cache
-
-type Bls12GrothParams = groth16::Parameters<Bls12>;
-type Bls12VerifyingKey = groth16::VerifyingKey<Bls12>;
-
-type Cache<G> = HashMap<String, Arc<G>>;
-type GrothMemCache = Cache<Bls12GrothParams>;
-type VerifyingKeyMemCache = Cache<Bls12VerifyingKey>;
-
-lazy_static! {
-    static ref GROTH_PARAM_MEMORY_CACHE: Mutex<GrothMemCache> = Default::default();
-    static ref VERIFYING_KEY_MEMORY_CACHE: Mutex<VerifyingKeyMemCache> = Default::default();
-}
-
-fn cache_lookup<F, G>(
-    cache_ref: &Mutex<Cache<G>>,
-    identifier: String,
-    generator: F,
-) -> error::Result<Arc<G>>
-where
-    F: FnOnce() -> error::Result<G>,
-    G: Send + Sync,
-{
-    info!(FCP_LOG, "trying parameters memory cache for: {}", &identifier; "target" => "params");
-    {
-        let cache = (*cache_ref).lock().unwrap();
-
-        if let Some(entry) = cache.get(&identifier) {
-            info!(FCP_LOG, "found params in memory cache for {}", &identifier; "target" => "params");
-            return Ok(entry.clone());
-        }
-    }
-
-    info!(FCP_LOG, "no params in memory cache for {}", &identifier; "target" => "params");
-
-    let new_entry = Arc::new(generator()?);
-    let res = new_entry.clone();
-    {
-        let cache = &mut (*cache_ref).lock().unwrap();
-        cache.insert(identifier, new_entry);
-    }
-
-    Ok(res)
-}
-
-#[inline]
-fn lookup_groth_params<F>(identifier: String, generator: F) -> error::Result<Arc<Bls12GrothParams>>
-where
-    F: FnOnce() -> error::Result<Bls12GrothParams>,
-{
-    cache_lookup(&*GROTH_PARAM_MEMORY_CACHE, identifier, generator)
-}
-
-#[inline]
-fn lookup_verifying_key<F>(
-    identifier: String,
-    generator: F,
-) -> error::Result<Arc<Bls12VerifyingKey>>
-where
-    F: FnOnce() -> error::Result<Bls12VerifyingKey>,
-{
-    let vk_identifier = format!("{}-verifying-key", &identifier);
-    cache_lookup(&*VERIFYING_KEY_MEMORY_CACHE, vk_identifier, generator)
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-fn get_zigzag_params(porep_config: PoRepConfig) -> error::Result<Arc<groth16::Parameters<Bls12>>> {
-    let public_params = public_params(
-        PaddedBytesAmount::from(porep_config),
-        usize::from(PoRepProofPartitions::from(porep_config)),
-    );
-
-    let get_params =
-        || ZigZagCompound::groth_params(&public_params, &ENGINE_PARAMS).map_err(Into::into);
-
-    Ok(lookup_groth_params(
-        format!(
-            "ZIGZAG[{}]",
-            usize::from(PaddedBytesAmount::from(porep_config))
-        ),
-        get_params,
-    )?)
-}
-
-fn get_post_params(post_config: PoStConfig) -> error::Result<Arc<groth16::Parameters<Bls12>>> {
-    let post_public_params = post_public_params(post_config);
-
-    let get_params = || {
-        <VDFPostCompound as CompoundProof<
-            Bls12,
-            VDFPoSt<PedersenHasher, Sloth>,
-            VDFPoStCircuit<Bls12>,
-        >>::groth_params(&post_public_params, &ENGINE_PARAMS)
-        .map_err(Into::into)
-    };
-
-    Ok(lookup_groth_params(
-        format!(
-            "POST[{}]",
-            usize::from(PaddedBytesAmount::from(post_config))
-        ),
-        get_params,
-    )?)
-}
-
-fn get_zigzag_verifying_key(porep_config: PoRepConfig) -> error::Result<Arc<Bls12VerifyingKey>> {
-    let public_params = public_params(
-        PaddedBytesAmount::from(porep_config),
-        usize::from(PoRepProofPartitions::from(porep_config)),
-    );
-
-    let get_verifying_key =
-        || ZigZagCompound::verifying_key(&public_params, &ENGINE_PARAMS).map_err(Into::into);
-
-    Ok(lookup_verifying_key(
-        format!(
-            "ZIGZAG[{}]",
-            usize::from(PaddedBytesAmount::from(porep_config))
-        ),
-        get_verifying_key,
-    )?)
-}
-
-fn get_post_verifying_key(post_config: PoStConfig) -> error::Result<Arc<Bls12VerifyingKey>> {
-    let post_public_params = post_public_params(post_config);
-
-    let get_verifying_key = || {
-        <VDFPostCompound as CompoundProof<
-            Bls12,
-            VDFPoSt<PedersenHasher, Sloth>,
-            VDFPoStCircuit<Bls12>,
-        >>::verifying_key(&post_public_params, &ENGINE_PARAMS)
-        .map_err(Into::into)
-    };
-
-    Ok(lookup_verifying_key(
-        format!(
-            "POST[{}]",
-            usize::from(PaddedBytesAmount::from(post_config))
-        ),
-        get_verifying_key,
-    )?)
-}
-
-const DEGREE: usize = 5;
-const EXPANSION_DEGREE: usize = 8;
-const SLOTH_ITER: usize = 0;
-const LAYERS: usize = 4; // TODO: 10;
-const TAPER_LAYERS: usize = 2; // TODO: 7
-const TAPER: f64 = 1.0 / 3.0;
-const POREP_MINIMUM_CHALLENGES: usize = 12; // FIXME: 8,000
-
-const DRG_SEED: [u32; 7] = [1, 2, 3, 4, 5, 6, 7]; // Arbitrary, need a theory for how to vary this over time.
-
-fn select_challenges(
-    partitions: usize,
-    minimum_total_challenges: usize,
-    layers: usize,
-    taper_layers: usize,
-    taper: f64,
-) -> LayerChallenges {
-    let mut count = 1;
-    let mut guess = LayerChallenges::Tapered {
-        count,
-        layers,
-        taper,
-        taper_layers,
-    };
-    while partitions * guess.total_challenges() < minimum_total_challenges {
-        count += 1;
-        guess = LayerChallenges::Tapered {
-            count,
-            layers,
-            taper,
-            taper_layers,
-        };
-    }
-    guess
-}
-
-fn setup_params(
-    sector_bytes: PaddedBytesAmount,
-    partitions: usize,
-) -> layered_drgporep::SetupParams {
-    let sector_bytes = usize::from(sector_bytes);
-
-    let challenges = select_challenges(
-        partitions,
-        POREP_MINIMUM_CHALLENGES,
-        LAYERS,
-        TAPER_LAYERS,
-        TAPER,
-    );
-
-    assert!(
-        sector_bytes % 32 == 0,
-        "sector_bytes ({}) must be a multiple of 32",
-        sector_bytes,
-    );
-    let nodes = sector_bytes / 32;
-    layered_drgporep::SetupParams {
-        drg: DrgParams {
-            nodes,
-            degree: DEGREE,
-            expansion_degree: EXPANSION_DEGREE,
-            seed: DRG_SEED,
-        },
-        sloth_iter: SLOTH_ITER,
-        layer_challenges: challenges,
-    }
-}
-
-pub fn public_params(
-    sector_bytes: PaddedBytesAmount,
-    partitions: usize,
-) -> layered_drgporep::PublicParams<DefaultTreeHasher, ZigZagBucketGraph<DefaultTreeHasher>> {
-    ZigZagDrgPoRep::<DefaultTreeHasher>::setup(&setup_params(sector_bytes, partitions)).unwrap()
-}
-
-type PostSetupParams = vdf_post::SetupParams<PedersenDomain, vdf_sloth::Sloth>;
-pub type PostPublicParams = vdf_post::PublicParams<PedersenDomain, vdf_sloth::Sloth>;
-
-const POST_CHALLENGE_COUNT: usize = 30;
-const POST_EPOCHS: usize = 3;
-pub const POST_SECTORS_COUNT: usize = 2;
-const POST_VDF_ROUNDS: usize = 1;
-
-lazy_static! {
-    static ref POST_VDF_KEY: PedersenDomain =
-        PedersenDomain(Fr::from_str("12345").unwrap().into_repr());
-}
-
-fn post_setup_params(post_config: PoStConfig) -> PostSetupParams {
-    let size = PaddedBytesAmount::from(post_config);
-
-    vdf_post::SetupParams::<PedersenDomain, vdf_sloth::Sloth> {
-        challenge_count: POST_CHALLENGE_COUNT,
-        sector_size: size.into(),
-        post_epochs: POST_EPOCHS,
-        setup_params_vdf: vdf_sloth::SetupParams {
-            key: *POST_VDF_KEY,
-            rounds: POST_VDF_ROUNDS,
-        },
-        sectors_count: POST_SECTORS_COUNT,
-    }
-}
-
-pub fn post_public_params(post_config: PoStConfig) -> PostPublicParams {
-    VDFPoSt::<PedersenHasher, vdf_sloth::Sloth>::setup(&post_setup_params(post_config)).unwrap()
-}
-
-fn commitment_from_fr<E: Engine>(fr: E::Fr) -> Commitment {
-    let mut commitment = [0; 32];
-    for (i, b) in fr_into_bytes::<E>(&fr).iter().enumerate() {
-        commitment[i] = *b;
-    }
-    commitment
-}
-
-fn pad_safe_fr(unpadded: &FrSafe) -> Fr32Ary {
-    let mut res = [0; 32];
-    res[0..31].copy_from_slice(unpadded);
-    res
+#[derive(Clone, Debug)]
+pub struct SealOutput {
+    pub comm_r: Commitment,
+    pub comm_r_star: Commitment,
+    pub comm_d: Commitment,
+    pub proof: Vec<u8>,
 }
 
 pub fn generate_post(
+    post_config: PoStConfig,
+    challenge_seed: ChallengeSeed,
+    input_parts: Vec<(Option<String>, Commitment)>,
+) -> error::Result<GeneratePoStDynamicSectorsCountOutput> {
+    return generate_post_dynamic(GeneratePoStDynamicSectorsCountInput {
+        post_config,
+        challenge_seed,
+        input_parts,
+    });
+}
+
+pub fn verify_post(
+    post_config: PoStConfig,
+    comm_rs: Vec<Commitment>,
+    challenge_seed: ChallengeSeed,
+    proofs: Vec<Vec<u8>>,
+    faults: Vec<u64>,
+) -> error::Result<VerifyPoStDynamicSectorsCountOutput> {
+    return verify_post_dynamic(VerifyPoStDynamicSectorsCountInput {
+        post_config,
+        comm_rs,
+        challenge_seed,
+        proofs,
+        faults,
+    });
+}
+
+pub fn seal<T: Into<PathBuf> + AsRef<Path>>(
+    porep_config: PoRepConfig,
+    in_path: T,
+    out_path: T,
+    prover_id_in: &FrSafe,
+    sector_id_in: &FrSafe,
+) -> error::Result<SealOutput> {
+    let sector_bytes = usize::from(PaddedBytesAmount::from(porep_config));
+
+    let mut cleanup = FileCleanup::new(&out_path);
+
+    // Copy unsealed data to output location, where it will be sealed in place.
+    copy(&in_path, &out_path)?;
+    let f_data = OpenOptions::new().read(true).write(true).open(&out_path)?;
+
+    // Zero-pad the data to the requested size by extending the underlying file if needed.
+    f_data.set_len(sector_bytes as u64)?;
+
+    let mut data = unsafe { MmapOptions::new().map_mut(&f_data).unwrap() };
+
+    // Zero-pad the prover_id to 32 bytes (and therefore Fr32).
+    let prover_id = pad_safe_fr(prover_id_in);
+    // Zero-pad the sector_id to 32 bytes (and therefore Fr32).
+    let sector_id = pad_safe_fr(sector_id_in);
+    let replica_id = replica_id::<DefaultTreeHasher>(prover_id, sector_id);
+
+    let compound_setup_params = compound_proof::SetupParams {
+        vanilla_params: &setup_params(
+            PaddedBytesAmount::from(porep_config),
+            usize::from(PoRepProofPartitions::from(porep_config)),
+        ),
+        engine_params: &(*ENGINE_PARAMS),
+        partitions: Some(usize::from(PoRepProofPartitions::from(porep_config))),
+    };
+
+    let compound_public_params = ZigZagCompound::setup(&compound_setup_params)?;
+
+    let (tau, aux) = ZigZagDrgPoRep::replicate(
+        &compound_public_params.vanilla_params,
+        &replica_id,
+        &mut data,
+        None,
+    )?;
+
+    // If we succeeded in replicating, flush the data and protect output from being cleaned up.
+    data.flush()?;
+    cleanup.success = true;
+
+    let public_tau = tau.simplify();
+
+    let public_inputs = layered_drgporep::PublicInputs {
+        replica_id,
+        tau: Some(public_tau),
+        comm_r_star: tau.comm_r_star,
+        k: None,
+    };
+
+    let private_inputs = layered_drgporep::PrivateInputs::<DefaultTreeHasher> {
+        aux,
+        tau: tau.layer_taus,
+    };
+
+    let groth_params = get_zigzag_params(porep_config)?;
+
+    info!(FCP_LOG, "got groth params ({}) while sealing", u64::from(PaddedBytesAmount::from(porep_config)); "target" => "params");
+
+    let proof = ZigZagCompound::prove(
+        &compound_public_params,
+        &public_inputs,
+        &private_inputs,
+        &groth_params,
+    )?;
+
+    let mut buf = Vec::with_capacity(
+        SINGLE_PARTITION_PROOF_LEN * usize::from(PoRepProofPartitions::from(porep_config)),
+    );
+
+    proof.write(&mut buf)?;
+
+    let comm_r = commitment_from_fr::<Bls12>(public_tau.comm_r.into());
+    let comm_d = commitment_from_fr::<Bls12>(public_tau.comm_d.into());
+    let comm_r_star = commitment_from_fr::<Bls12>(tau.comm_r_star.into());
+
+    // Verification is cheap when parameters are cached,
+    // and it is never correct to return a proof which does not verify.
+    verify_seal(
+        porep_config,
+        comm_r,
+        comm_d,
+        comm_r_star,
+        prover_id_in,
+        sector_id_in,
+        &buf,
+    )
+    .expect("post-seal verification sanity check failed");
+
+    Ok(SealOutput {
+        comm_r,
+        comm_r_star,
+        comm_d,
+        proof: buf,
+    })
+}
+
+pub fn verify_seal(
+    porep_config: PoRepConfig,
+    comm_r: Commitment,
+    comm_d: Commitment,
+    comm_r_star: Commitment,
+    prover_id_in: &FrSafe,
+    sector_id_in: &FrSafe,
+    proof_vec: &[u8],
+) -> error::Result<bool> {
+    let sector_bytes = PaddedBytesAmount::from(porep_config);
+    let prover_id = pad_safe_fr(prover_id_in);
+    let sector_id = pad_safe_fr(sector_id_in);
+    let replica_id = replica_id::<DefaultTreeHasher>(prover_id, sector_id);
+
+    let comm_r = bytes_into_fr::<Bls12>(&comm_r)?;
+    let comm_d = bytes_into_fr::<Bls12>(&comm_d)?;
+    let comm_r_star = bytes_into_fr::<Bls12>(&comm_r_star)?;
+
+    let compound_setup_params = compound_proof::SetupParams {
+        vanilla_params: &setup_params(
+            PaddedBytesAmount::from(porep_config),
+            usize::from(PoRepProofPartitions::from(porep_config)),
+        ),
+        engine_params: &(*ENGINE_PARAMS),
+        partitions: Some(usize::from(PoRepProofPartitions::from(porep_config))),
+    };
+
+    let compound_public_params: compound_proof::PublicParams<
+        '_,
+        Bls12,
+        ZigZagDrgPoRep<'_, DefaultTreeHasher>,
+    > = ZigZagCompound::setup(&compound_setup_params)?;
+
+    let public_inputs = layered_drgporep::PublicInputs::<<DefaultTreeHasher as Hasher>::Domain> {
+        replica_id,
+        tau: Some(Tau {
+            comm_r: comm_r.into(),
+            comm_d: comm_d.into(),
+        }),
+        comm_r_star: comm_r_star.into(),
+        k: None,
+    };
+
+    let verifying_key = get_zigzag_verifying_key(porep_config)?;
+
+    info!(FCP_LOG, "got verifying key ({}) while verifying seal", u64::from(sector_bytes); "target" => "params");
+
+    let proof = MultiProof::new_from_reader(
+        Some(usize::from(PoRepProofPartitions::from(porep_config))),
+        proof_vec,
+        &verifying_key,
+    )?;
+
+    ZigZagCompound::verify(
+        &compound_public_params,
+        &public_inputs,
+        &proof,
+        &ChallengeRequirements {
+            minimum_challenges: POREP_MINIMUM_CHALLENGES,
+        },
+    )
+    .map_err(Into::into)
+}
+
+pub fn get_unsealed_range<T: Into<PathBuf> + AsRef<Path>>(
+    porep_config: PoRepConfig,
+    sealed_path: T,
+    output_path: T,
+    prover_id_in: &FrSafe,
+    sector_id_in: &FrSafe,
+    offset: UnpaddedByteIndex,
+    num_bytes: UnpaddedBytesAmount,
+) -> error::Result<(UnpaddedBytesAmount)> {
+    let prover_id = pad_safe_fr(prover_id_in);
+    let sector_id = pad_safe_fr(sector_id_in);
+    let replica_id = replica_id::<DefaultTreeHasher>(prover_id, sector_id);
+
+    let f_in = File::open(sealed_path)?;
+    let mut data = Vec::new();
+    f_in.take(u64::from(PaddedBytesAmount::from(porep_config)))
+        .read_to_end(&mut data)?;
+
+    let f_out = File::create(output_path)?;
+    let mut buf_writer = BufWriter::new(f_out);
+
+    let unsealed = ZigZagDrgPoRep::extract_all(
+        &public_params(
+            PaddedBytesAmount::from(porep_config),
+            usize::from(PoRepProofPartitions::from(porep_config)),
+        ),
+        &replica_id,
+        &data,
+    )?;
+
+    let written = write_unpadded(&unsealed, &mut buf_writer, offset.into(), num_bytes.into())?;
+
+    Ok(UnpaddedBytesAmount(written as u64))
+}
+
+fn verify_post_dynamic(
+    dynamic: VerifyPoStDynamicSectorsCountInput,
+) -> error::Result<VerifyPoStDynamicSectorsCountOutput> {
+    let fixed = verify_post_spread_input(dynamic)?
+        .iter()
+        .map(verify_post_fixed_sectors_count)
+        .collect();
+
+    verify_post_collect_output(fixed)
+}
+
+fn generate_post_dynamic(
     dynamic: GeneratePoStDynamicSectorsCountInput,
 ) -> error::Result<GeneratePoStDynamicSectorsCountOutput> {
     let n = { dynamic.input_parts.len() };
@@ -330,18 +312,7 @@ pub fn generate_post(
     generate_post_collect_output(n, fixed_output)
 }
 
-pub fn verify_post(
-    dynamic: VerifyPoStDynamicSectorsCountInput,
-) -> error::Result<VerifyPoStDynamicSectorsCountOutput> {
-    let fixed = verify_post_spread_input(dynamic)?
-        .iter()
-        .map(verify_post_fixed_sectors_count)
-        .collect();
-
-    verify_post_collect_output(fixed)
-}
-
-pub fn generate_post_fixed_sectors_count(
+fn generate_post_fixed_sectors_count(
     fixed: &GeneratePoStFixedSectorsCountInput,
 ) -> error::Result<GeneratePoStFixedSectorsCountOutput> {
     let faults: Vec<u64> = Vec::new();
@@ -471,7 +442,6 @@ fn verify_post_fixed_sectors_count(
     Ok(VerifyPoStFixedSectorsCountOutput { is_valid })
 }
 
-type Tree = MerkleTree<PedersenDomain, <PedersenHasher as Hasher>::Function>;
 fn make_merkle_tree<T: Into<PathBuf> + AsRef<Path>>(
     sealed_path: T,
     bytes: PaddedBytesAmount,
@@ -483,250 +453,22 @@ fn make_merkle_tree<T: Into<PathBuf> + AsRef<Path>>(
     public_params(bytes, 1).graph.merkle_tree(&data)
 }
 
-#[derive(Clone, Debug)]
-pub struct SealOutput {
-    pub comm_r: Commitment,
-    pub comm_r_star: Commitment,
-    pub comm_d: Commitment,
-    pub proof: Vec<u8>,
-}
-
-/// Minimal support for cleaning (deleting) a file unless it was successfully populated.
-struct FileCleanup<T: AsRef<Path>> {
-    path: T,
-    success: bool,
-}
-
-impl<'a, T: AsRef<Path>> FileCleanup<T> {
-    fn new(path: T) -> FileCleanup<T> {
-        FileCleanup {
-            path,
-            success: false,
-        }
+fn commitment_from_fr<E: Engine>(fr: E::Fr) -> Commitment {
+    let mut commitment = [0; 32];
+    for (i, b) in fr_into_bytes::<E>(&fr).iter().enumerate() {
+        commitment[i] = *b;
     }
+    commitment
 }
 
-impl<T: AsRef<Path>> Drop for FileCleanup<T> {
-    fn drop(&mut self) {
-        if !self.success {
-            let _ = remove_file(&self.path);
-        }
-    }
-}
-
-pub fn seal<T: Into<PathBuf> + AsRef<Path>>(
-    porep_config: PoRepConfig,
-    in_path: T,
-    out_path: T,
-    prover_id_in: &FrSafe,
-    sector_id_in: &FrSafe,
-) -> error::Result<SealOutput> {
-    let sector_bytes = usize::from(PaddedBytesAmount::from(porep_config));
-
-    let mut cleanup = FileCleanup::new(&out_path);
-
-    // Copy unsealed data to output location, where it will be sealed in place.
-    copy(&in_path, &out_path)?;
-    let f_data = OpenOptions::new().read(true).write(true).open(&out_path)?;
-
-    // Zero-pad the data to the requested size by extending the underlying file if needed.
-    f_data.set_len(sector_bytes as u64)?;
-
-    let mut data = unsafe { MmapOptions::new().map_mut(&f_data).unwrap() };
-
-    // Zero-pad the prover_id to 32 bytes (and therefore Fr32).
-    let prover_id = pad_safe_fr(prover_id_in);
-    // Zero-pad the sector_id to 32 bytes (and therefore Fr32).
-    let sector_id = pad_safe_fr(sector_id_in);
-    let replica_id = replica_id::<DefaultTreeHasher>(prover_id, sector_id);
-
-    let compound_setup_params = compound_proof::SetupParams {
-        vanilla_params: &setup_params(
-            PaddedBytesAmount::from(porep_config),
-            usize::from(PoRepProofPartitions::from(porep_config)),
-        ),
-        engine_params: &(*ENGINE_PARAMS),
-        partitions: Some(usize::from(PoRepProofPartitions::from(porep_config))),
-    };
-
-    let compound_public_params = ZigZagCompound::setup(&compound_setup_params)?;
-
-    let (tau, aux) = ZigZagDrgPoRep::replicate(
-        &compound_public_params.vanilla_params,
-        &replica_id,
-        &mut data,
-        None,
-    )?;
-
-    // If we succeeded in replicating, flush the data and protect output from being cleaned up.
-    data.flush()?;
-    cleanup.success = true;
-
-    let public_tau = tau.simplify();
-
-    let public_inputs = layered_drgporep::PublicInputs {
-        replica_id,
-        tau: Some(public_tau),
-        comm_r_star: tau.comm_r_star,
-        k: None,
-    };
-
-    let private_inputs = layered_drgporep::PrivateInputs::<DefaultTreeHasher> {
-        aux,
-        tau: tau.layer_taus,
-    };
-
-    let groth_params = get_zigzag_params(porep_config)?;
-
-    info!(FCP_LOG, "got groth params ({}) while sealing", u64::from(PaddedBytesAmount::from(porep_config)); "target" => "params");
-
-    let proof = ZigZagCompound::prove(
-        &compound_public_params,
-        &public_inputs,
-        &private_inputs,
-        &groth_params,
-    )?;
-
-    let mut buf = Vec::with_capacity(
-        SINGLE_PARTITION_PROOF_LEN * usize::from(PoRepProofPartitions::from(porep_config)),
-    );
-
-    proof.write(&mut buf)?;
-
-    let comm_r = commitment_from_fr::<Bls12>(public_tau.comm_r.into());
-    let comm_d = commitment_from_fr::<Bls12>(public_tau.comm_d.into());
-    let comm_r_star = commitment_from_fr::<Bls12>(tau.comm_r_star.into());
-
-    // Verification is cheap when parameters are cached,
-    // and it is never correct to return a proof which does not verify.
-    verify_seal(
-        porep_config,
-        comm_r,
-        comm_d,
-        comm_r_star,
-        prover_id_in,
-        sector_id_in,
-        &buf,
-    )
-    .expect("post-seal verification sanity check failed");
-
-    Ok(SealOutput {
-        comm_r,
-        comm_r_star,
-        comm_d,
-        proof: buf,
-    })
-}
-
-pub fn get_unsealed_range<T: Into<PathBuf> + AsRef<Path>>(
-    porep_config: PoRepConfig,
-    sealed_path: T,
-    output_path: T,
-    prover_id_in: &FrSafe,
-    sector_id_in: &FrSafe,
-    offset: UnpaddedByteIndex,
-    num_bytes: UnpaddedBytesAmount,
-) -> error::Result<(UnpaddedBytesAmount)> {
-    let prover_id = pad_safe_fr(prover_id_in);
-    let sector_id = pad_safe_fr(sector_id_in);
-    let replica_id = replica_id::<DefaultTreeHasher>(prover_id, sector_id);
-
-    let f_in = File::open(sealed_path)?;
-    let mut data = Vec::new();
-    f_in.take(u64::from(PaddedBytesAmount::from(porep_config)))
-        .read_to_end(&mut data)?;
-
-    let f_out = File::create(output_path)?;
-    let mut buf_writer = BufWriter::new(f_out);
-
-    let unsealed = ZigZagDrgPoRep::extract_all(
-        &public_params(
-            PaddedBytesAmount::from(porep_config),
-            usize::from(PoRepProofPartitions::from(porep_config)),
-        ),
-        &replica_id,
-        &data,
-    )?;
-
-    let written = write_unpadded(&unsealed, &mut buf_writer, offset.into(), num_bytes.into())?;
-
-    Ok(UnpaddedBytesAmount(written as u64))
-}
-
-pub fn verify_seal(
-    porep_config: PoRepConfig,
-    comm_r: Commitment,
-    comm_d: Commitment,
-    comm_r_star: Commitment,
-    prover_id_in: &FrSafe,
-    sector_id_in: &FrSafe,
-    proof_vec: &[u8],
-) -> error::Result<bool> {
-    let sector_bytes = PaddedBytesAmount::from(porep_config);
-    let prover_id = pad_safe_fr(prover_id_in);
-    let sector_id = pad_safe_fr(sector_id_in);
-    let replica_id = replica_id::<DefaultTreeHasher>(prover_id, sector_id);
-
-    let comm_r = bytes_into_fr::<Bls12>(&comm_r)?;
-    let comm_d = bytes_into_fr::<Bls12>(&comm_d)?;
-    let comm_r_star = bytes_into_fr::<Bls12>(&comm_r_star)?;
-
-    let compound_setup_params = compound_proof::SetupParams {
-        vanilla_params: &setup_params(
-            PaddedBytesAmount::from(porep_config),
-            usize::from(PoRepProofPartitions::from(porep_config)),
-        ),
-        engine_params: &(*ENGINE_PARAMS),
-        partitions: Some(usize::from(PoRepProofPartitions::from(porep_config))),
-    };
-
-    let compound_public_params: compound_proof::PublicParams<
-        '_,
-        Bls12,
-        ZigZagDrgPoRep<'_, DefaultTreeHasher>,
-    > = ZigZagCompound::setup(&compound_setup_params)?;
-
-    let public_inputs = layered_drgporep::PublicInputs::<<DefaultTreeHasher as Hasher>::Domain> {
-        replica_id,
-        tau: Some(Tau {
-            comm_r: comm_r.into(),
-            comm_d: comm_d.into(),
-        }),
-        comm_r_star: comm_r_star.into(),
-        k: None,
-    };
-
-    let verifying_key = get_zigzag_verifying_key(porep_config)?;
-
-    info!(FCP_LOG, "got verifying key ({}) while verifying seal", u64::from(sector_bytes); "target" => "params");
-
-    let proof = MultiProof::new_from_reader(
-        Some(usize::from(PoRepProofPartitions::from(porep_config))),
-        proof_vec,
-        &verifying_key,
-    )?;
-
-    ZigZagCompound::verify(
-        &compound_public_params,
-        &public_inputs,
-        &proof,
-        &ChallengeRequirements {
-            minimum_challenges: POREP_MINIMUM_CHALLENGES,
-        },
-    )
-    .map_err(Into::into)
+fn pad_safe_fr(unpadded: &FrSafe) -> Fr32Ary {
+    let mut res = [0; 32];
+    res[0..31].copy_from_slice(unpadded);
+    res
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-
-    use rand::{thread_rng, Rng};
-    use sector_base::api::disk_backed_storage::new_sector_store;
-    use sector_base::api::disk_backed_storage::TEST_SECTOR_SIZE;
-    use sector_base::api::sector_class::SectorClass;
-    use sector_base::api::sector_size::SectorSize;
-    use sector_base::api::sector_store::SectorStore;
     use std::fs::create_dir_all;
     use std::fs::File;
     use std::io::Read;
@@ -734,7 +476,17 @@ mod tests {
     use std::io::SeekFrom;
     use std::io::Write;
     use std::thread;
+
+    use rand::{thread_rng, Rng};
     use tempfile::NamedTempFile;
+
+    use sector_base::api::disk_backed_storage::new_sector_store;
+    use sector_base::api::disk_backed_storage::TEST_SECTOR_SIZE;
+    use sector_base::api::sector_class::SectorClass;
+    use sector_base::api::sector_size::SectorSize;
+    use sector_base::api::sector_store::SectorStore;
+
+    use super::*;
 
     const TEST_CLASS: SectorClass = SectorClass(
         SectorSize(TEST_SECTOR_SIZE),
@@ -924,23 +676,23 @@ mod tests {
         let comm_rs = vec![comm_r, comm_r];
         let challenge_seed = rng.gen();
 
-        let post_output = generate_post(GeneratePoStDynamicSectorsCountInput {
-            post_config: h.store.proofs_config().post_config(),
+        let post_output = generate_post(
+            h.store.proofs_config().post_config(),
             challenge_seed,
-            input_parts: vec![
+            vec![
                 (Some(h.sealed_access.clone()), comm_r),
                 (Some(h.sealed_access.clone()), comm_r),
             ],
-        })
+        )
         .expect("PoSt generation failed");
 
-        let result = verify_post(VerifyPoStDynamicSectorsCountInput {
-            post_config: h.store.proofs_config().post_config(),
+        let result = verify_post(
+            h.store.proofs_config().post_config(),
             comm_rs,
             challenge_seed,
-            proofs: post_output.proofs,
-            faults: post_output.faults,
-        })
+            post_output.proofs,
+            post_output.faults,
+        )
         .expect("failed to run verify_post");
 
         assert!(result.is_valid, "verification of valid proof failed");
@@ -1184,25 +936,5 @@ mod tests {
     #[ignore]
     fn post_verify_test() {
         post_verify_aux(TEST_CLASS, BytesAmount::Max);
-    }
-
-    #[test]
-    fn partition_layer_challenges_test() {
-        let f = |partitions| {
-            select_challenges(
-                partitions,
-                POREP_MINIMUM_CHALLENGES,
-                LAYERS,
-                TAPER_LAYERS,
-                TAPER,
-            )
-            .all_challenges()
-        };
-        // Update to ensure all supported PoRepProofPartitions options are represented here.
-        assert_eq!(vec![1, 1, 2, 2], f(usize::from(PoRepProofPartitions(2))));
-
-        assert_eq!(vec![3, 3, 4, 5], f(1));
-        assert_eq!(vec![1, 1, 2, 2], f(2));
-        assert_eq!(vec![1, 1, 1, 1], f(4));
     }
 }
